requirements.txt
Flask==2.0.1
Flask-Limiter==2.4.0
MarkupSafe==2.0.1
Jinja2==3.0.3
Flask-WTF==1.1.1
python-dotenv
cryptography
markdown
pymdown-extensions